requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.3
Django==5.0.6
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-crontab==0.7.1
django-debug-toolbar==4.3.0
django-filter==24.2
django-redis==5.4.0
django-timezone-field==6.1.0
djangorestframework==3.15.1
dnspython==2.6.1
drf-yasg==1.21.7
eventlet==0.36.1
greenlet==3.0.3
idna==3.7
inflection==0.5.1
kombu==5.3.7
packaging==24.0
postgres==4.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
psycopg2-pool==1.2
python-crontab==3.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.4
requests==2.31.0
setuptools==69.5.1
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13

jsonrpcserver~=5.0.9